+++ packages/typespec-rust/test/spector/azure/resource-manager/resources/tests/resources_nested_client_test.rs
@@ -4,9 +4,19 @@
 
 mod common;
 
-use azure_core::time::OffsetDateTime;
+use azure_core::{
+    http::{
+        poller::{PollerOptions, PollerStatus, StatusMonitor},
+        RequestContent, StatusCode,
+    },
+    time::{Duration, OffsetDateTime},
+};
 use futures::StreamExt;
-use spector_armresources::models::{NestedProxyResource, NestedProxyResourceProperties};
+use spector_armresources::models::{
+    CreatedByType, NestedProxyResource, NestedProxyResourceProperties, ProvisioningState,
+    ResourcesNestedClientCreateOrReplaceOptions, ResourcesNestedClientDeleteOptions,
+    ResourcesNestedClientUpdateOptions,
+};
 use time::{Date, Month, Time};
 
 #[tokio::test]
@@ -158,7 +168,7 @@ fn get_valid_nested_resource() -> NestedProxyResource {
         name: Some("nested".to_string()),
         type_prop: Some("Azure.ResourceManager.Resources/topLevelTrackedResources/top/nestedProxyResources".to_string()),
         properties: Some(NestedProxyResourceProperties {
-            provisioning_state: Some(spector_armresources::models::ProvisioningState::Succeeded),
+            provisioning_state: Some(ProvisioningState::Succeeded),
             description: Some("valid".to_string()),
         }),
         // Using from_json to create the system_data since it's marked as #[non_exhaustive]
@@ -187,4 +197,267 @@ fn validate_timestamps(
     // Verify date components match expected values
     assert_eq!(created_at, Some(expected_dt));
     assert_eq!(last_modified_at, Some(expected_dt));
+}
+
+#[tokio::test]
+async fn create_or_replace() {
+    let client = common::create_client().get_resources_nested_client();
+
+    let create_or_replace_request: RequestContent<NestedProxyResource> = NestedProxyResource {
+        properties: Some(NestedProxyResourceProperties {
+            description: Some("valid".to_string()),
+            ..Default::default()
+        }),
+        ..Default::default()
+    }
+    .try_into()
+    .unwrap();
+
+    let options = Some(ResourcesNestedClientCreateOrReplaceOptions {
+        method_options: PollerOptions {
+            frequency: Duration::seconds(1),
+            ..Default::default()
+        },
+    });
+
+    let mut poller = client
+        .create_or_replace(
+            "test-rg",
+            "top",
+            "nested",
+            create_or_replace_request.clone(),
+            options.clone(),
+        )
+        .unwrap();
+
+    let mut poll_count = 0;
+    while let Some(result) = poller.next().await {
+        poll_count += 1;
+        let response = result.unwrap();
+        let http_status = response.status();
+        let status_monitor = response.into_model().unwrap();
+        let poller_status = status_monitor.status();
+        match poll_count {
+            1 => {
+                assert_eq!(http_status, StatusCode::Ok);
+                assert_eq!(poller_status, PollerStatus::Succeeded);
+            }
+            _ => {
+                panic!("unexpected poll count");
+            }
+        }
+    }
+    assert_eq!(poll_count, 1);
+
+    let poller = client
+        .create_or_replace(
+            "test-rg",
+            "top",
+            "nested",
+            create_or_replace_request,
+            options,
+        )
+        .unwrap();
+    let final_result = poller.await.unwrap().into_model().unwrap();
+
+    assert_eq!(final_result.id, Some("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top/nestedProxyResources/nested".to_string()));
+    assert_eq!(final_result.name, Some("nested".to_string()));
+    assert_eq!(
+        final_result.type_prop,
+        Some(
+            "Azure.ResourceManager.Resources/topLevelTrackedResources/top/nestedProxyResources"
+                .to_string()
+        )
+    );
+
+    assert!(final_result.properties.is_some());
+    assert_eq!(
+        final_result.properties.as_ref().unwrap().provisioning_state,
+        Some(ProvisioningState::Succeeded)
+    );
+    assert_eq!(
+        final_result.properties.as_ref().unwrap().description,
+        Some("valid".to_string())
+    );
+
+    assert!(final_result.system_data.is_some());
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_by,
+        Some("AzureSDK".to_string())
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_by_type,
+        Some(CreatedByType::User)
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_at,
+        Some(OffsetDateTime::new_utc(
+            Date::from_calendar_date(2024, Month::October, 4).unwrap(),
+            Time::from_hms_milli(0, 56, 7, 442).unwrap(),
+        ))
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().last_modified_by,
+        Some("AzureSDK".to_string())
+    );
+    assert_eq!(
+        final_result
+            .system_data
+            .as_ref()
+            .unwrap()
+            .last_modified_by_type,
+        Some(CreatedByType::User)
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().last_modified_at,
+        Some(OffsetDateTime::new_utc(
+            Date::from_calendar_date(2024, Month::October, 4).unwrap(),
+            Time::from_hms_milli(0, 56, 7, 442).unwrap(),
+        ))
+    );
+}
+
+#[tokio::test]
+async fn delete() {
+    let client = common::create_client().get_resources_nested_client();
+
+    let options = Some(ResourcesNestedClientDeleteOptions {
+        method_options: PollerOptions {
+            frequency: Duration::seconds(1),
+            ..Default::default()
+        },
+    });
+
+    let mut poller = client.delete("test-rg", "top", "nested", options).unwrap();
+
+    let mut poll_count = 0;
+    while let Some(result) = poller.next().await {
+        poll_count += 1;
+        let response = result.unwrap();
+        let http_status = response.status();
+        let status_monitor = response.into_model().unwrap();
+        let poller_status = status_monitor.status();
+        match poll_count {
+            1 => {
+                assert_eq!(http_status, StatusCode::NoContent);
+                assert_eq!(poller_status, PollerStatus::Succeeded);
+            }
+            _ => {
+                panic!("unexpected poll count");
+            }
+        }
+    }
+    assert_eq!(poll_count, 1);
+}
+
+#[tokio::test]
+async fn update() {
+    let client = common::create_client().get_resources_nested_client();
+
+    let update_request: RequestContent<NestedProxyResource> = NestedProxyResource {
+        properties: Some(NestedProxyResourceProperties {
+            description: Some("valid2".to_string()),
+            ..Default::default()
+        }),
+        ..Default::default()
+    }
+    .try_into()
+    .unwrap();
+
+    let options = Some(ResourcesNestedClientUpdateOptions {
+        method_options: PollerOptions {
+            frequency: Duration::seconds(1),
+            ..Default::default()
+        },
+    });
+
+    let mut poller = client
+        .update(
+            "test-rg",
+            "top",
+            "nested",
+            update_request.clone(),
+            options.clone(),
+        )
+        .unwrap();
+
+    let mut poll_count = 0;
+    while let Some(result) = poller.next().await {
+        poll_count += 1;
+        let response = result.unwrap();
+        let http_status = response.status();
+        let status_monitor = response.into_model().unwrap();
+        let poller_status = status_monitor.status();
+        match poll_count {
+            1 => {
+                assert_eq!(http_status, StatusCode::Ok);
+                assert_eq!(poller_status, PollerStatus::Succeeded);
+            }
+            _ => {
+                panic!("unexpected poll count");
+            }
+        }
+    }
+    assert_eq!(poll_count, 1);
+
+    let poller = client
+        .update("test-rg", "top", "nested", update_request, options)
+        .unwrap();
+    let final_result = poller.await.unwrap().into_model().unwrap();
+
+    assert_eq!(final_result.id, Some("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/topLevelTrackedResources/top/nestedProxyResources/nested".to_string()));
+    assert_eq!(final_result.name, Some("nested".to_string()));
+    assert_eq!(
+        final_result.type_prop,
+        Some(
+            "Azure.ResourceManager.Resources/topLevelTrackedResources/top/nestedProxyResources"
+                .to_string()
+        )
+    );
+
+    assert!(final_result.properties.is_some());
+    assert_eq!(
+        final_result.properties.as_ref().unwrap().provisioning_state,
+        Some(ProvisioningState::Succeeded)
+    );
+    assert_eq!(
+        final_result.properties.as_ref().unwrap().description,
+        Some("valid2".to_string())
+    );
+
+    assert!(final_result.system_data.is_some());
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_by,
+        Some("AzureSDK".to_string())
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_by_type,
+        Some(CreatedByType::User)
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().created_at,
+        Some(OffsetDateTime::new_utc(
+            Date::from_calendar_date(2024, Month::October, 4).unwrap(),
+            Time::from_hms_milli(0, 56, 7, 442).unwrap(),
+        ))
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().last_modified_by,
+        Some("AzureSDK".to_string())
+    );
+    assert_eq!(
+        final_result
+            .system_data
+            .as_ref()
+            .unwrap()
+            .last_modified_by_type,
+        Some(CreatedByType::User)
+    );
+    assert_eq!(
+        final_result.system_data.as_ref().unwrap().last_modified_at,
+        Some(OffsetDateTime::new_utc(
+            Date::from_calendar_date(2024, Month::October, 4).unwrap(),
+            Time::from_hms_milli(0, 56, 7, 442).unwrap(),
+        ))
+    );
 }